Modify build system

+++ build/xref.sh
@@ -1,26 +1,35 @@
 #
-# Appelle phpxref en se mettant dans le repertoire d'install de phpxref.
-# Si on l'appelle a partir d'un autre repertoire, le resultat n'inclut pas les
-# feuilles de style.
-# Ce wrapper permet de ne pas mettre le repertoire en dur dans le fichier de
-# config
+# Set default directory to phpxref install directory and, then, launch phpxref.
+# If phpxref is called from another directory, the result does not include
+# the style sheets (phpxref bug).
+# This wrapper script allows not to set the output path in the configuration file.
+#
 # Variables :
-# INPUT = Repertoire des sources (relatif par rapport au repertoire du script)
-# OUTPUT = repertoire resultat
-# PHPXREF_DIR = repertoire d'install de phpxref
+# INPUT = Directory containing the source files (relative to the directory
+#         containing the script)
+# OUTPUT = Output dir
+# PHPXREF_DIR = phpxref install dir
+#
+# $1 = Config file
+
+if [ -z "$PHPXREF_DIR" ] ; then
+	echo "PHPXREF_DIR is not defined"
+	exit 1
+fi
 
 TMP=/tmp/.t$$
 /bin/rm -rf $TMP
 
-cd `dirname $0`
-dir=`/bin/pwd`
+save_wd=`pwd`
+cd $INPUT
+SOURCE=`pwd`
+cd $save_wd
 
-SOURCE=$INPUT
-echo $SOURCE | grep '^/' >/dev/null 2>&1 || SOURCE=$dir/$INPUT
-
-echo "SOURCE=$SOURCE" >$TMP
-echo "OUTPUT=$OUTPUT" >>$TMP
-grep -v '^SOURCE=' <$1 | grep -v 'OUTPUT=' >>$TMP
+(
+echo "SOURCE=$SOURCE"
+echo "OUTPUT=$OUTPUT"
+grep -v '^SOURCE=' <$1 | grep -v 'OUTPUT='
+) >$TMP
 
 cd $PHPXREF_DIR
 perl phpxref.pl -c $TMP
